Only run CI in either pull requests or master
This is useful because if you are a project collaborator and do
not fork to push changes to a new branch then CI will run for both
the pull request and the branch to which you're pushing your
changes, duplicating all jobs.

I think it's acceptable to disable CI for any other branch than
master and allow it if a branch has an associated pull request.

One can create a draft pull request to develop along with CI in
a local branch.