Bump dev-dependencies for examples
Updates the dependencies used by examples to current - some have had API-breaking changes, so some examples are updated to catch up with those as well.

Note that the hyper-server example doesn't seem to work when tested, but it also didn't work before this PR. Looked at it a bit and found that poll_accept() returns Pending via incoming.poll_next() and then poll_accept() isn't invoked again. hyper-client meanwhile works fine, so the issue doesn't seem to be related to the AsyncRead updates.