docs: Add an ARCHITECTURE.md document
It would be useful to have an ARCHITECTURE document that can be used by
new contributors to get a "bird's eye view" of how things are arranged
in smol. This commit adds a basic ARCHITECTURE.md document with the aim
that it can be used by prospective contributors.

So far only the lower-level items of the ecosystem have been described.

Signed-off-by: John Nunley <dev@notgull.net>